Fixed post scrambling
- Closes #33
- The issue above was actually caused by the parent container. Elements
are recycled by some controls. This way, existing SparklrText controls
that were out of view  are appended and come back into view with other
data on the other side. For instance: If you scroll down in a ListBox,
elements that are disappearing on the top will come back into view from
below. The SparklrText Control updates images asynchronously. This way,
if you scroll fast enough, the image loading will finish when the
element is already recycled. This commit contains some checks to make
sure that the loaded image is still the one assigned to the post.